refactor: spend -> expense 로 변경
