+++ snapsplit/src/features/trip/[tripId]/budget/expense/_components/expense-form/CategorySection.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import CategoryItem from './category/CategoryItem';
-import { SPEND_CATEGORY } from '@/shared/constants/spend';
+import { EXPENSE_CATEGORY } from '@/shared/constants/expense';
 import { useState } from 'react';
 
 export default function CategorySection() {
@@ -10,7 +10,7 @@ export default function CategorySection() {
     <div className="flex flex-col items-start w-full gap-3">
       <div className="text-body-3">지출 카테고리</div>
       <div className="flex items-center gap-2 w-full">
-        {SPEND_CATEGORY.map((category) => (
+        {EXPENSE_CATEGORY.map((category) => (
           <CategoryItem key={category.name} name={category.name} selectedCategory={selectedCategory} setSelectedCategory={() => setSelectedCategory(category.name)} />
         ))}
       </div>
