+++ snapsplit/src/features/trip/[tripId]/budget/expense/_components/expense-form/DescriptionSection.tsx
