+++ snapsplit/src/features/trip/[tripId]/budget/expense/_components/expense-form/payment/PaymentRow.tsx
@@ -20,7 +20,7 @@ export default function PaymentRow({ name }: Props) {
       <div className="flex items-center">
         <div className="flex items-center justify-center w-20">
           <button onClick={toggleCheck} className={`flex items-center justify-center w-6 h-6 rounded-full ${isChecked ? 'bg-primary text-white' : 'border-[1px] border-grey-250'}`}>
-            <Image src={isChecked ? "/svg/check-spend-white.svg" : "/svg/check-spend-grey.svg"} alt="" width={24} height={24} />
+            <Image src={isChecked ? "/svg/check-expense-white.svg" : "/svg/check-expense-grey.svg"} alt="" width={24} height={24} />
           </button>
         </div>
         <input type="text" className="flex items-center justify-center w-20 text-center text-grey-450" placeholder="0" />
